mediexplain_ai/requirements.txt
streamlit
pytesseract
openai
Pillow
gtts
opencv-python
